# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Home Logbook API
  version: 0.11.0
servers:
- url: http://192.168.68.62:8787/api/v1
security:
- bearerAuth: []
paths:
...
  /event-types/{key}/latest-event:
    parameters:
    - in: path
      name: key
      required: true
      schema:
        type: string
    get:
      summary: Get the most recent event for an event type
      description: Uses the event type key and returns the newest event ordered by startedAt, then id.
      responses:
        '200':
          description: Most recent event
        '404':
          description: Unknown event type or no recorded events
    patch:
      summary: Update the most recent event for an event type
      description: Updates only supplied start-time, value, unit, text-value, note, or metadata fields. Numeric
        values are validated and normalized by the backend. For point events, changing startedAt also changes endedAt
        to the same timestamp. For duration events, the existing end time is preserved.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LatestEventUpdate'
      responses:
        '200':
          description: Updated most recent event
        '400':
          description: Invalid value
          unit: null
          or start time: null
        '404':
          description: Unknown event type or no recorded events
        '409':
          description: The selected latest event changed concurrently
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
  /events:
    get:
      summary: List and search events
      description: Returns events in reverse chronological order with numbered pagination. The note filter performs
        a case-insensitive substring search. It currently uses a plain case-insensitive substring scan; a trigram
        index can be added later without changing this API contract.
      parameters:
      - in: query
        name: eventTypeId
        schema:
          type: string
          format: uuid
      - in: query
        name: eventTypeKey
        schema:
          type: string
      - in: query
        name: from
        schema:
          type: string
          format: date-time
      - in: query
        name: to
        schema:
          type: string
          format: date-time
      - in: query
        name: before
        schema:
          type: string
          format: date-time
      - in: query
        name: ongoing
        schema:
          type: boolean
      - in: query
        name: note
        description: Case-insensitive text contained in the event note.
        schema:
          type: string
          minLength: 1
          maxLength: 500
      - in: query
        name: page
        schema:
          type: integer
          minimum: 1
          default: 1
      - in: query
        name: pageSize
        schema:
          type: integer
          minimum: 1
          maximum: 100
          default: 100
      - in: query
        name: limit
        deprecated: true
        description: Legacy alias for pageSize; cannot be combined with pageSize.
        schema:
          type: integer
          minimum: 1
          maximum: 500
      responses:
        '200':
          description: Paginated events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedEvents'
        '400':
          description: Invalid filters or pagination values
  /events/log:
    post:
      summary: Log a point-in-time event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PointEventLog'
      responses:
        '201':
          description: Logged point event
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
  /events/start:
    post:
      summary: Start a duration event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DurationEventStart'
      responses:
        '201':
          description: Started duration event
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
  /events/end:
    post:
      summary: End one ongoing duration event
      description: Supply an eventId, or an eventTypeId/eventTypeKey to end the newest ongoing duration for that
        type. Optionally supply value and unitKey to set or replace the event's numeric value as part of the same
        completion operation. Numeric values are normalized by the backend; omitting value preserves the existing
        value.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventEnd'
      responses:
        '200':
          description: Ended event
        '400':
          description: endedAt is earlier than the event start
        '404':
          description: Event
          event type: null
          or ongoing event not found: null
        '409':
          description: Event is a point event
          already ended: null
          or was ended concurrently: null
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
  /events/{id}:
    parameters:
    - in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid
    get:
      summary: Get an event
      responses:
        '200':
          description: Event
    patch:
      summary: Edit an individual event
      description: Updates value/unit, text value, note, start time, or end time. Numeric values are normalized
        by the backend. Point events must retain identical start and end times. Set endedAt to null to make a duration
        event ongoing.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventUpdate'
      responses:
        '200':
          description: Updated event
        '400':
          description: Invalid value
          unit: null
          or timestamp shape: null
        '404':
          description: Event not found
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
    delete:
      summary: Delete an event
      responses:
        '204':
          description: Deleted
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    EventValue:
      type: object
      properties:
        value:
          type:
          - number
          - 'null'
          description: Original input value
        unitKey:
          type:
          - string
          - 'null'
          description: Exact compatible input unit
        textValue:
          type:
          - string
          - 'null'
        note:
          type:
          - string
          - 'null'
        metadata:
          type: object
          additionalProperties: true
    LatestEventUpdate:
      allOf:
      - type: object
        properties:
          startedAt:
            type: string
            format: date-time
      - $ref: '#/components/schemas/EventValue'
      minProperties: 1
      description: Supply at least one editable field. unitKey may only be supplied with a numeric value. Changing
        a point event's startedAt moves its endedAt to the same instant; a duration event keeps its current endedAt.
    PointEventLog:
      allOf:
      - type: object
        properties:
          eventTypeId:
            type: string
            format: uuid
          eventTypeKey:
            type: string
          occurredAt:
            type: string
            format: date-time
      - $ref: '#/components/schemas/EventValue'
    DurationEventStart:
      allOf:
      - type: object
        properties:
          eventTypeId:
            type: string
            format: uuid
          eventTypeKey:
            type: string
          startedAt:
            type: string
            format: date-time
      - $ref: '#/components/schemas/EventValue'
    EventUpdate:
      type: object
      minProperties: 1
      properties:
        startedAt:
          type: string
          format: date-time
        endedAt:
          type:
          - string
          - 'null'
          format: date-time
        value:
          type:
          - number
          - 'null'
        unitKey:
          type:
          - string
          - 'null'
        textValue:
          type:
          - string
          - 'null'
          maxLength: 4000
        note:
          type:
          - string
          - 'null'
          maxLength: 4000
        metadata:
          type: object
          additionalProperties: true
    EventPagination:
      type: object
      required:
      - page
      - pageSize
      - total
      - totalPages
      - hasPrevious
      - hasNext
      properties:
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 1
        hasPrevious:
          type: boolean
        hasNext:
          type: boolean
    PaginatedEvents:
      type: object
      required:
      - events
      - pagination
      properties:
        events:
          type: array
          items:
            type: object
        pagination:
          $ref: '#/components/schemas/EventPagination'
    EventEnd:
      type: object
      description: Provide exactly one of eventId, eventTypeId, or eventTypeKey.
      oneOf:
      - required:
        - eventId
      - required:
        - eventTypeId
      - required:
        - eventTypeKey
      properties:
        eventId:
          type: string
          format: uuid
        eventTypeId:
          type: string
          format: uuid
        eventTypeKey:
          type: string
        endedAt:
          type: string
          format: date-time
        value:
          type:
          - number
          - 'null'
          description: Optional original input value to store while ending
        unitKey:
          type:
          - string
          - 'null'
          description: Optional compatible unit; requires value
  parameters:
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: false
      description: Optional unique key for safely retrying POST, PATCH, and DELETE requests. Reusing a key with
        the same request returns the stored response; reusing it with a different request returns 409.
      schema:
        type: string
        minLength: 1
        maxLength: 200